migration edited

+++ database/migrations/2024_09_10_094736_create_storage_items_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up()
 {
-    Schema::create('storageitems', function (Blueprint $table) {
-        $table->id(); // Primärschlüssel
-        $table->unsignedBigInteger('user_id');
-        $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        $table->unsignedBigInteger('product_id'); // Verknüpfung zu Produkten
-        $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade'); // Fremdschlüssel hinzufügen
-        $table->string('product_name'); // Name des Produkts
-        $table->float('quantity'); // Menge des Produkts im Lager
-        $table->unsignedBigInteger('unit_id'); // Verknüpfung zur Einheit
-        $table->timestamps(); // Erstellt created_at und updated_at
-    });
+    if (!Schema::hasTable('storageitems')) {
+        Schema::create('storageitems', function (Blueprint $table) {
+            $table->unsignedBigInteger('id')->autoIncrement();
+            $table->unsignedBigInteger('user_id');
+            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            $table->unsignedBigInteger('product_id');
+            $table->string('product_name');
+            $table->float('quantity', 53);
+            $table->unsignedBigInteger('unit_id');
+            $table->timestamps();
+        });
+    }
 }
+
